% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{MDEI-package}
\alias{MDEI-package}
\title{MDEI}
\description{
MDEI estimates  effects.
}
\author{
\strong{Maintainer}: Marc Ratkovic \email{ratkovic@princeton.edu}

Authors:
\itemize{
  \item Nithin Kavi \email{nithinkavi@college.harvard.edu}
}

Other contributors:
\itemize{
  \item Dustin Tingley \email{dtingley@gov.harvard.edu} [contributor]
}

}
