\name{NithinPackage-package}
\alias{NithinPackage-package}
\alias{NithinPackage}
\docType{package}
\title{
\packageTitle{NithinPackage}
}
\description{
\packageDescription{NithinPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NithinPackage}
\packageIndices{NithinPackage}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{NithinPackage}

Maintainer: \packageMaintainer{NithinPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
