% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDEI_summary_plot.R
\name{plot.MDEI}
\alias{plot.MDEI}
\title{Plot an MDEI Object}
\usage{
\method{plot}{MDEI}(
  obj,
  xvar = "treat",
  sigval = 0,
  target = "tau",
  colors = c(gray(0.7), gray(0)),
  cex.point = 0.5,
  xlabel = "",
  ylabel = "",
  ...
)
}
\arguments{
\item{obj}{An object of class MDEI.}

\item{...}{Additional values to be passed to \code{summary}.}

\item{nvars}{Label for y-axis of figure.  Default is \code{""}.}
}
\description{
Plots an object of class MDEI
}
