% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDEI_summary_plot.R
\name{summary.MDEI}
\alias{summary.MDEI}
\title{Summary of an MDEI Object}
\usage{
\method{summary}{MDEI}(object, features = 10, ...)
}
\arguments{
\item{object}{An object of class MDEI.}

\item{features}{Number of spline bases to include.}

\item{...}{Additional arguments to be passed to \code{plot}.}
}
\value{
\describe{
\item{coeftable}{A table with three columns: the names of selected spline interactions, the average coefficient, and
proportion of time it was included in the model. Averages over taken over subsamples in the
split sample strategy.Note that the coefficients are interactions between
spline interactions that can be accessed through \code{obj$internal$Xmat.spline}.}
}
}
\description{
Summary of an object of class MDEI.
}
