% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseregTE_auxiliary.R
\name{MDEI}
\alias{MDEI}
\title{MDEI function}
\usage{
MDEI(y, treat, X, splits = 10, alpha = 0.9)
}
\arguments{
\item{y}{The outcome variable, a vector.}

\item{treat}{The treatment variable, a vector.}

\item{X}{A matrix of covariates.}

\item{splits}{Number of repeated cross-fitting steps to implement.}

\item{alpha}{The desired level of the confidence band.}
}
\description{
Implements the Method of Direct Estimation and Inference
}
