% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDEI.R
\name{MDEI}
\alias{MDEI}
\title{MDEI function}
\usage{
MDEI(
  y,
  treat,
  X,
  splits = 10,
  alpha = 0.9,
  samplesplit = TRUE,
  conformal = TRUE
)
}
\arguments{
\item{y}{The outcome variable, a vector.}

\item{treat}{The treatment variable, a vector.}

\item{X}{A matrix of covariates.}

\item{splits}{Number of repeated cross-fitting steps to implement.}

\item{alpha}{The desired level of the confidence band.}

\item{samplesplit}{Whether to use a sample splitting approach. Default is \code{TRUE}.}

\item{conformal}{Whether to generate a conformal bands or use a critical value from the
normal approximation.  Default is \code{TRUE}.}
}
\value{
\describe{
\item{tau.est}{The estimated marginal effect.}
\item{CIs.tau}{Conformal Confiden.}
\item{critical.values}{Conformal critical values.}
\item{Ey.x}{Mean of outcome given only covariates.}
\item{coefficients}{The list of all nonparameteric bases and the proportion of sample splits that they were selected.}
\item{internal}{Internal objects used for development and diagnostics.}
}
}
\description{
Implements the Method of Direct Estimation and Inference
}
\examples{
n<-200

X <- matrix(rnorm(n*3), nrow = n)
treat <- rnorm(n)
y <- treat^2 + X[,2] + rnorm(n)

set.seed(1)
m1 <- MDEI(y, treat, X, splits=5, alpha=.9)

# Accuracy
cor(m1$tau.est, treat*2)
cor(m1$theta.est, treat^2)

# Coverage
mean(apply(m1$CIs.tau-2*treat,1,prod)<0)
}
\references{
Ratkovic, Marc and Dustin Tingley.  2023. "Estimation and Inference on Nonlinear and Heterogeneous Effects."  The Journal of Politics.
}
